put v2 methods into their respective modules and remove v2 file
