+++ lib/mauth/remote_authenticator.rb
@@ -20,6 +20,25 @@ module MAuth
         make_mauth_request(authentication_ticket)
       end
 
+      def signature_valid_v2!(object)
+        unless object.is_a?(MAuth::Request)
+          msg = "Remote Authenticator can only authenticate requests; received #{object.inspect}"
+          raise ArgumentError, msg
+        end
+
+        authentication_ticket = {
+          verb: object.attributes_for_signing[:verb],
+          app_uuid: object.signature_app_uuid,
+          client_signature: object.signature,
+          request_url: object.attributes_for_signing[:request_url],
+          request_time: object.mcc_time,
+          b64encoded_body: Base64.encode64(object.attributes_for_signing[:body] || ''),
+          query_string: object.attributes_for_signing[:query_string],
+          token: object.signature_token
+        }
+        make_mauth_request(authentication_ticket)
+      end
+
       def make_mauth_request(authentication_ticket)
         begin
           response = mauth_connection.post("/mauth/#{mauth_api_version}/authentication_tickets.json", 'authentication_ticket' => authentication_ticket)
